Main.java

+++ MolinaPineda_01092025/src/MolinaPineda_01092025/Main.java
@@ -89,7 +89,7 @@ public class Main {
             
         }
         //CALCULO DE VALORES
-        promedio = suma /total_elementos;
+        promedio = (double) suma /total_elementos;
         
         //SE IMPRIME LA SALIDA DEL PROGRAMA AL USUARIO CON LOS VALORES ANTERIORMENTE CALCULADOS
         System.out.println(suma+" es la sumatoria total de elementos, la cantidad de elementos digitados es de " +total_elementos);
